+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 Here are the latest changes to the site. View the [commit log](https://github.com/vegeta897/snow-stamp/commits/main) for a full list of changes.
 
+## 2021-12-07
+
+- ☑️ Concluded visitor survey - [see the results](assets/survey-2021-12-07.md)
+
 ## 2021-12-03
 
 - 🔍 Added Snow-Stamp search support in address bar <details><summary>Show me!</summary>![Searching Snow-Stamp in address bar](assets/snow-search.gif?raw=true)</details>
@@ -13,7 +17,7 @@ Here are the latest changes to the site. View the [commit log](https://github.co
 
 ## 2021-11-18
 
-- 📋 Added link to [user survey](https://forms.gle/6QZgX1mZQNwZsaJP7)
+- 📋 Added link to visitor survey
 
 ## 2021-10-24